+++ theme/js/components/dataset/search-result.vue
@@ -2,13 +2,17 @@
 Vue. -->
 
 <template>
-  <article class="dataset-card dataset-search-result">
+  <article class="dataset-card dataset-search-result py-xs">
     <div class="card-logo" v-if="organization">
       <Placeholder
         type="dataset"
         :src="organization.logo_thumbnail"
         :alt="organization.name"
       />
+      <div class="logo-badge logo-badge--bottom-right">
+        <span v-html="lock" v-if="private" />
+        <span v-html="certified" v-else-if="organizationCertified" />
+      </div>
     </div>
     <div class="card-logo" v-else-if="owner">
       <Placeholder
@@ -16,10 +20,6 @@ Vue. -->
         :src="owner.logo_thumbnail"
         :alt="owner.fullname"
       />
-      <div class="logo-badge">
-        <span v-html="lock" v-if="private" />
-        <span v-html="certified" v-else-if="organization?.public_service" />
-      </div>
     </div>
     <div class="card-logo" v-else>
       <Placeholder type="dataset" />
@@ -39,9 +39,9 @@ Vue. -->
         <dt>{{ $t("Frequency") }}</dt>
         <dd>{{ frequency }}</dd>
       </div>
-      <div v-if="geozone">
+      <div v-if="geoZone">
         <dt>{{ $t("Spatial coverage") }}</dt>
-        <dd>{{ geozone.join(", ") }}</dd>
+        <dd>{{ geoZone.join(", ") }}</dd>
       </div>
       <div v-if="spatial?.granularity">
         <dt>{{ $t("Territorial coverage granularity") }}</dt>
@@ -69,11 +69,12 @@ Vue. -->
 import Placeholder from "../utils/placeholder";
 import certified from "svg/certified.svg";
 import lock from "svg/private.svg";
+import useOrganizationCertified from "../../composables/useOrganizationCertified";
+import useGeoZone from "../../composables/useGeoZone";
 
 export default {
   props: {
     title: String,
-    image_url: String,
     organization: Object,
     owner: Object,
     description: String,
@@ -87,27 +88,15 @@ export default {
   components: {
     Placeholder,
   },
-  data() {
+  setup(props) {
+    const {organizationCertified} = useOrganizationCertified(props.organization)
+    const {geoZone} = useGeoZone(props.spatial)
     return {
-      geozone: null,
+      certified,
+      lock,
+      organizationCertified,
+      geoZone,
     };
-  },
-  async mounted() {
-    this.certified = certified;
-    this.lock = lock;
-    //Fetching geozone names on load (they're not included in the dataset object)
-
-    const zones = this?.spatial?.zones;
-    if (zones) {
-      let promises = zones.map((zone) =>
-        this.$api
-          .get("spatial/zone/" + zone)
-          .then((resp) => resp.data)
-          .then((obj) => obj && obj?.properties?.name)
-      );
-
-      this.geozone = await Promise.all(promises);
-    }
-  },
+  }
 };
 </script>