+++ theme/js/plugins/api.js
@@ -7,10 +7,10 @@ export const generateCancelToken = () => CancelToken.source();
 // Instantiate axios with base URL from config
 // No need for CSRF or anything fancy here
 // TODO : maybe add interceptor to better handle errors ?
-const api = axios.create({
+export const api = axios.create({
   baseURL: config.api_root,
 });
-const apiv2 = axios.create({
+export const apiv2 = axios.create({
   baseURL: config.api_2_root,
 });